% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINSEG.R
\name{BINSEG}
\alias{BINSEG}
\title{Detect a series of change points with binary segmentation}
\usage{
BINSEG(data, segs)
}
\arguments{
\item{data}{a vector of data}

\item{segs}{the number of segments}
}
\value{
a data matrix of the change points and costs
}
\description{
Detect a series of change points with binary segmentation and return the points and costs
}
\examples{
Rfunc2::BINSEG(c(1,2,3,4,5,6),2)
}
