% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGEPOINT.R
\name{CHANGEPOINT}
\alias{CHANGEPOINT}
\title{Detect a changepoint}
\usage{
CHANGEPOINT(data)
}
\arguments{
\item{data}{a vector of data}
}
\value{
the change point location and the reduction in cost
}
\description{
Detect a Changepoint in a vector using sum of squares as a cost
}
\examples{
Rfunc2::CHANGEPOINT(c(1,2,3,4,5,6))
}
